+++ .gitignore
@@ -3,3 +3,4 @@
 /key.json
 /.env
 /.idea
+/vendor
