--- src/judge/Makefile
@@ -1,2 +1,0 @@
-judge:judge.go
-	go build --race -o ./../..
